ENERGY_DATA_ETRACTION/requirements.txt
pymongo==4.5.0
pandas==2.0.3
python-dotenv==1.0.0